backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1

# Validation & Config
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Formula engine (safe eval)
simpleeval

# Utils
python-dotenv==1.0.1
